Add in columns and fix migration
