+++ alembic/versions/de5276fad9ef_record_signer_dod_id.py
@@ -1,0 +1,34 @@
+"""Record signer DOD ID
+
+Revision ID: de5276fad9ef
+Revises: 1f690989e38e
+Create Date: 2019-02-11 10:51:51.419346
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = 'de5276fad9ef'
+down_revision = '1f690989e38e'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('task_orders', sa.Column('level_of_warrant', sa.Numeric(scale=2), nullable=True))
+    op.add_column('task_orders', sa.Column('signed_at', sa.DateTime(), nullable=True))
+    op.add_column('task_orders', sa.Column('signer_dod_id', sa.String(), nullable=True))
+    op.add_column('task_orders', sa.Column('unlimited_level_of_warrant', sa.Boolean(), nullable=True))
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('task_orders', 'unlimited_level_of_warrant')
+    op.drop_column('task_orders', 'signer_dod_id')
+    op.drop_column('task_orders', 'signed_at')
+    op.drop_column('task_orders', 'level_of_warrant')
+    # ### end Alembic commands ###
