+++ templates/portfolios/blank_slate.html
@@ -1,0 +1,19 @@
+{% extends "base.html" %}
+
+{% from "components/empty_state.html" import EmptyState %}
+{% from "components/tooltip.html" import Tooltip %}
+
+{% block global_sidenav %}
+{% endblock %}
+
+{% block content %}
+  {{
+    EmptyState(
+      action_href=url_for("task_orders.new", screen=1),
+      action_label=("portfolios.index.empty.start_button" | translate),
+      icon="cloud",
+      message=("portfolios.index.empty.title" | translate),
+    )
+  }}
+{% endblock %}
+
